13_docker/03_flask-postgres-app/webapp/requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.9
gunicorn==22.0.0
